requirements.txt
customtkinter
Pillow
windnd
matplotlib
requests
